also rename test modules
